telegram_bot/requirements.txt
python-telegram-bot==21.6
requests
fastapi
uvicorn